refactor: extract bitmap downsampling to LawnchairUtils
Address review feedback: the inSampleSize calculation was duplicated by
the new carousel code and SearchTargetFactory.FilesTarget. Move it into
LawnchairUtils as calculateInSampleSize() plus a decodeSampledBitmapFromFile()
helper, and reuse it from both call sites.
